chore: fetch data after adding tag
